Move DependencyRequirements
